test: isolate RP2040 crystal via repro
